Move serialization to core.
