+++ source/GlassView.Core/Extensions.cs
@@ -1,0 +1,13 @@
+using System.Text.Json;
+using GlassView.Core.Serialization;
+
+namespace GlassView.Core;
+
+public static class Extensions
+{
+    public static JsonSerializerOptions EnableGlassView(this JsonSerializerOptions options)
+    {
+        options.Converters.Add(new BenchmarkSerializer());
+        return options;
+    }
+}
